webapi/webapi-sensors-w3c-tests/sensors/Accelerometer_reading.https.html

<!DOCTYPE html>
<!--
Copyright (c) 2016 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-->

<meta charset="utf-8">
<title>Accelerometer Test: reading</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="author" title="Xiuqi Jiang" href="mailto:xiuqix.jiang@intel.com">
<link rel="help" href="https://w3c.github.io/accelerometer/">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id="log"></div>
<script>

let sensor, sensor1, sensor2, cachedReading1, cachedReading2;

setup(() => {
  sensor = new Accelerometer();
  sensor1 = new Accelerometer();
  sensor2 = new Accelerometer();
  sensor1.start();
  sensor2.start();
});

async_test(t => {
  let sensor = new Accelerometer();
  sensor.start();
  sensor.onactivate = t.step_func_done(() => {
    let cachedReading = sensor.reading;
    let cachedXaxis = cachedReading.x;
    sensor.stop();
    assert_equals(cachedReading.x, cachedXaxis);
  });
  sensor.onerror = t.step_func_done(event => {
    assert_unreached(event.error.name + ":" + event.error.message);
  });
}, "Test that sensor reading must be immutable.");

async_test(t => {
  sensor1.onactivate = t.step_func_done(() => {
    cachedReading1 = sensor1.reading;
    cachedReading2 = sensor2.reading;
    //both sensors share the same reading instance
    assert_equals(cachedReading1, cachedReading2);
    //after first sensor stops its reading is null, second sensor remains
    sensor1.stop();
    assert_equals(sensor1.reading, null);
    assert_equals(String(sensor2.reading), "[object AccelerometerReading]");
  });
  sensor1.onerror = t.step_func_done(event => {
    assert_unreached(event.error.name + ":" + event.error.message);
  });
}, "Test that sensor reading is correct.");

async_test(t => {
  t.step_timeout(() => {
    sensor2.onchange = t.step_func_done(() => {
      let cachedReading3 = sensor2.reading;
      assert_not_equals(cachedReading2, cachedReading3);
      sensor2.stop();
    });
    sensor2.onerror = t.step_func_done(event => {
      assert_unreached(event.error.name + ":" + event.error.message);
    });
  }, 1000);
}, "Test that the sensor reading is updated when time passes.");

</script>
